Feature: Allow free subscriptions to be created on Stripe
We realized that we want to be able to create free subscriptions, and therefore
there we should not need a credit card number to be passed in order to do so.
Previously, we were getting the customers `default_card` from Stripe and calling
`last4` on it, regardless of the amount of the plan.

Stripe does allow a null credit card for free subscriptions, however because of
the above implementation, if we created a free subscription we would still be
asking stripe for the `default_card` and calling `last4` and blowing up!

Now we do a simple check to see the plan amount for the `subscription` and if it
is `zero?`, we do not ask for the `default_card` and the
`last_four_of_credit_card` will be equal to nil.

Note: We now need to change the database structure and validations to allow
`last_four_of_credit_card` to be null on the `subscription` model.