Feature: Allow payment methods to be updated for existing subscriptions
There was previously no way to pass in a payment token for an existing
subscription.

Now, we update the subscription with the passed in payment token.  We do this
whether or not the subscription was just now created in that call.  This is
a bit less performant and if we need to change it later, we can, but for right
now, we're good with this.

We also rescue from an `InvalidRequestError` in the case that we attempt to
update the subscription with a token that we've already used in the past.
Mainly this happens when we create a new subscription and then attempt to update
the subscription with that same payment token.